- Minor fixed / improvements

+++ runningdinner-backend/src/main/resources/db/migration/V2.9__AddResendToMessageTask.sql
@@ -1,5 +1,5 @@
 ALTER TABLE runningdinner.MessageTask ADD COLUMN originalSender TEXT;
 
-ALTER TABLE runningdinner.MessageTask ADD COLUMN resendCount int8;
+ALTER TABLE runningdinner.MessageTask ADD COLUMN resendCount integer;
 UPDATE runningdinner.MessageTask SET resendCount = 0;
 ALTER TABLE runningdinner.MessageTask ALTER COLUMN resendCount SET NOT NULL;
